move schema processor
